Kody-Camera, Maze A* Pathfinding
New Features:
-RTS style camera implemented, move backwards and forwards by pushing to
the edges of the screen, zoom with scroll wheel and press alt to rotate.
-Created the right maze based off of the assignment specification.
Obstacles need to be rescaled so the corridors are the right size.
-Adapted the pathfinding from lab 8 so that an object can navigate
through the maze. Click on the maze to tell the object to go there.
-Cleaned up the scene creation, rendering and input functions. For any
functionality that you add try and keep them in your own cpp files and
declare them inside DemoApp.h. This way we can just call the rendering
function for that aspect rather than cluttering up the main rendering
function (and others like input and scene setup).